+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add api to get hourly prices for multiple ptids in columnar format
 - Unify the api for price information to be 'isone/da', 'isone/rt' from 'dalmp', 'rtlmp'.  
   Will deprecate the existing routes in 2024-01-27
+- Add a rtlmp client for this data similar to the existing dalmp client
 
 ## Release 2023-01-24
 - Fix CT rates download.  I was downloading only the first page!  Duh.